+++ toot/api.py
@@ -76,11 +76,14 @@ def request_access_token(app, authorization_code):
     return http.process_response(response).json()
 
 
-def post_status(app, user, status, visibility='public', media_ids=None):
+def post_status(app, user, status, visibility='public', media_ids=None,
+                sensitive=False, spoiler_text=None):
     return http.post(app, user, '/api/v1/statuses', {
         'status': status,
         'media_ids[]': media_ids,
         'visibility': visibility,
+        'sensitive': sensitive,
+        'spoiler_text': spoiler_text,
     }).json()